feat: refactor help logic into cli_help.py
